test(controller): move ensure step tests into ensure package
- move step-specific tests beside their implementations
- retain reconciliation and envtest coverage in the controller package
- remove the legacy controller test adapter
- centralize shared fixtures in package-local test helpers
- clean up stale helpers, comments, and scaffold TODOs
